translation_system/persistence_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiomysql==0.2.0
PyMySQL==1.1.0
PyYAML==6.0.1
python-dateutil==2.8.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==0.21.0